+++ app/src/main/java/com/bintina/goouttolunchmvvm/login/model/database/UserDao.kt
@@ -11,5 +11,5 @@ interface UserDao {
     fun createUser(user: User)
 
     @Query("SELECT * FROM User WHERE userId = :userId")
-    fun getUser(userId: Int): LiveData<User>
+    fun getUser(userId: Long): LiveData<User>
 }
